Add teensy_uC_programmer.ino with modified SRAM transfer protocol
The SRAM transfer functionality is broken on the last iteration of the teensy_uC_programmer. The UART over USB chip on the Teensy does not function well with raw binary transmissions. Occasional carriage return \r and line feed \n bytes trigger the Serial API on the Teensy spuriously/randomly.
Furthermore, the UART FIFO buffer is limited to only 64 bytes on the Teensy 3.6.

This rewrite of the teensy uC programmer application introduces a robust protocol with preamble and packet length markers that are used in an integrity handshake with the sending Python application.

The SRAM binary is broken up into packets and now are consistently exchanged. Finally, a separate CRC check is calculated on the SRAM contents after the transaction and compared with the CRC calculation on the origin system.